[DROOLS-7174] Minor cleanup
