+++ efesto/efesto-core/efesto-runtime-manager/efesto-runtime-manager-core/src/main/java/org/kie/efesto/runtimemanager/core/service/RuntimeManagerImpl.java
@@ -18,35 +18,15 @@ package org.kie.efesto.runtimemanager.core.service;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
 
-import org.kie.efesto.common.api.cache.EfestoClassKey;
-import org.kie.efesto.common.api.cache.EfestoIdentifierClassKey;
 import org.kie.efesto.runtimemanager.api.model.EfestoInput;
 import org.kie.efesto.runtimemanager.api.model.EfestoOutput;
 import org.kie.efesto.runtimemanager.api.model.EfestoRuntimeContext;
-import org.kie.efesto.runtimemanager.api.service.KieRuntimeService;
 import org.kie.efesto.runtimemanager.api.service.RuntimeManager;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import static org.kie.efesto.runtimemanager.api.utils.SPIUtils.getDiscoveredKieRuntimeServices;
-import static org.kie.efesto.runtimemanager.api.utils.SPIUtils.getKieRuntimeService;
-import static org.kie.efesto.runtimemanager.api.utils.SPIUtils.getKieRuntimeServiceFromEfestoRuntimeContext;
+import static org.kie.efesto.runtimemanager.core.service.RuntimeManagerUtils.getOptionalOutput;
 
 public class RuntimeManagerImpl implements RuntimeManager {
-
-    private static final Logger logger = LoggerFactory.getLogger(RuntimeManagerImpl.class.getName());
-
-    protected static final Map<EfestoClassKey, List<KieRuntimeService>> firstLevelCache = new HashMap<>();
-    protected static final Map<EfestoIdentifierClassKey, KieRuntimeService> secondLevelCache = new HashMap<>();
-
-    static {
-        populateFirstLevelCache(firstLevelCache);
-    }
 
     @Override
     public Collection<EfestoOutput> evaluateInput(EfestoRuntimeContext context, EfestoInput... toEvaluate) {
@@ -60,70 +40,5 @@ public class RuntimeManagerImpl implements RuntimeManager {
         return toReturn;
     }
 
-    static void populateFirstLevelCache(final Map<EfestoClassKey, List<KieRuntimeService>> toPopulate) {
-        List<KieRuntimeService> discoveredKieRuntimeServices = getDiscoveredKieRuntimeServices();
-        populateFirstLevelCache(discoveredKieRuntimeServices, toPopulate);
-    }
-
-    static void populateFirstLevelCache(final List<KieRuntimeService> discoveredKieRuntimeServices,
-                                        final Map<EfestoClassKey, List<KieRuntimeService>> toPopulate) {
-        discoveredKieRuntimeServices.forEach(kieRuntimeService -> {
-            EfestoClassKey efestoClassKey = kieRuntimeService.getEfestoClassKeyIdentifier();
-            toPopulate.merge(efestoClassKey, List.of(kieRuntimeService), (previous,
-                                                                          toAdd) -> {
-                List<KieRuntimeService> toReturn = new ArrayList<>();
-                toReturn.addAll(previous);
-                toReturn.addAll(toAdd);
-                return toReturn;
-            });
-        });
-    }
-
-    static Optional<EfestoOutput> getOptionalOutput(EfestoRuntimeContext context, EfestoInput input) {
-        Optional<KieRuntimeService> retrieved = getKieRuntimeServiceLocal(context, input);
-        return retrieved.isPresent() ? retrieved.flatMap(kieRuntimeService -> kieRuntimeService.evaluateInput(input,
-                                                                                                              context)) : Optional.empty();
-    }
-
-    static Optional<KieRuntimeService> getKieRuntimeServiceLocal(EfestoRuntimeContext context, EfestoInput input) {
-
-        KieRuntimeService cachedKieRuntimeService = getKieRuntimeServiceFromSecondLevelCache(input);
-        if (cachedKieRuntimeService != null) {
-            return Optional.of(cachedKieRuntimeService);
-        }
-
-        Optional<KieRuntimeService> retrieved = getKieRuntimeServiceFromFirstLevelCache(context, input);
-        if (retrieved.isEmpty()) {
-            logger.warn("Cannot find KieRuntimeService for {}, looking inside context classloader",
-                        input.getModelLocalUriId());
-            retrieved = getKieRuntimeServiceFromEfestoRuntimeContext(input, context);
-            if (retrieved.isPresent()) {
-                KieRuntimeService toAdd = retrieved.get();
-                List<KieRuntimeService> stored = firstLevelCache.get(input.getFirstLevelCacheKey());
-                if (stored == null) {
-                    stored = new ArrayList<>();
-                    firstLevelCache.put(input.getFirstLevelCacheKey(), stored);
-                }
-                stored.add(toAdd);
-            }
-        }
-        if (retrieved.isEmpty()) {
-            logger.warn("Cannot find KieRuntimeService for {}", input.getModelLocalUriId());
-        } else {
-            secondLevelCache.put(input.getSecondLevelCacheKey(), retrieved.get());
-        }
-        return retrieved;
-    }
-
-    static KieRuntimeService getKieRuntimeServiceFromSecondLevelCache(EfestoInput input) {
-        return secondLevelCache.get(input.getSecondLevelCacheKey());
-    }
-
-    static Optional<KieRuntimeService> getKieRuntimeServiceFromFirstLevelCache(EfestoRuntimeContext context, EfestoInput input) {
-        List<KieRuntimeService> discoveredServices = firstLevelCache.get(input.getFirstLevelCacheKey());
-        return (discoveredServices != null && !discoveredServices.isEmpty()) ?
-                getKieRuntimeService(discoveredServices, input, context) :
-                Optional.empty();
-    }
 
 }
